groupnorm fp16

+++ TensorRT-8.5.3/plugin/groupNormPlugin/groupNormPlugin.cpp
@@ -211,7 +211,7 @@ int32_t GroupNormPlugin::enqueue(PluginTensorDesc const* inputDesc, PluginTensor
         mParams.cPerBlock = cPerBlock;
         mParams.cPerGroup = mParams.c / mParams.groups;
         mParams.hwc = mParams.hw * mParams.c;
-        mParams.invHWC = 1.F / (float) (mParams.hw * mParams.cPerGroup);
+        mParams.invHWC = __float2half(1.F / (float) (mParams.hw * mParams.cPerGroup));
         mParams.groupsPerBlock = cPerBlock / mParams.cPerGroup;
 
         cudaMemsetAsync(mParams.redBuffer, 0, getWorkspaceSizeInBytes(), stream);
